backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
spacy==3.7.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.4.0
pandas==2.2.2
numpy==1.25.2
Pillow==10.1.0
reportlab==4.1.3
pdfminer.six==20240715
pytesseract==0.4.3